refactor: consolidate drawer-specific utilities into component file
Move usePrefersReducedMotion, usePresence hooks and DOM utility functions into $drawer.tsx as they are only used by this component.
