--- app/hooks/usePrefersReducedMotion.ts
@@ -1,26 +1,0 @@
-import { useEffect, useState } from "hono/jsx"
-
-const MEDIA_QUERY = "(prefers-reduced-motion: reduce)"
-
-export const usePrefersReducedMotion = () => {
-  const [prefersReducedMotion, setPrefersReducedMotion] = useState(false)
-
-  useEffect(() => {
-    if (typeof window === "undefined" || !window.matchMedia) return
-    const mediaQuery = window.matchMedia(MEDIA_QUERY)
-    setPrefersReducedMotion(mediaQuery.matches)
-
-    const handleChange = (event: MediaQueryListEvent) =>
-      setPrefersReducedMotion(event.matches)
-
-    if (typeof mediaQuery.addEventListener === "function") {
-      mediaQuery.addEventListener("change", handleChange)
-      return () => mediaQuery.removeEventListener("change", handleChange)
-    }
-
-    mediaQuery.addListener(handleChange)
-    return () => mediaQuery.removeListener(handleChange)
-  }, [])
-
-  return prefersReducedMotion
-}
